TrustlessSidechain.Governance.MultiSig

Description

Implement a security mechanism that requires n out of m signatures on the transaction.

Documentation

data MultiSigGovParams

Parameters of the security mechanism. Note that setting requiredSignatures to a value greater than `length governanceMembers` will result in governance that can never approve anything.

NOTE: the order of entries in the governanceMembers matters! Since MultiSigGovParams is used to parameterize the multi-signature governance minting policy, changing the order of elements will change the hash of the policy.

Constructors

MultiSigGovParams

Fields

governanceMembers :: [PubKeyHash]

Members of the governance

requiredSignatures :: Integer

Minimal required number of signatures

serialisableGovernanceMultiSigPolicy :: SerialisedScript