TrustlessSidechain.PlutusPrelude

Documentation

class Semigroup a where

Plutus Tx version of Semigroup.

Methods

(<>) :: a -> a -> a infixr 6

Plutus Tx version of (<>).

data Either a b

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

Expand

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Constructors

Left a

Right b

data Bool

Constructors

False

True

data Ordering

Constructors

LT

EQ

GT

data Maybe a

The Maybe type encapsulates an optional value. A value of type Maybe a either contains a value of type a (represented as Just a), or it is empty (represented as Nothing). Using Maybe is a good way to deal with errors or exceptional cases without resorting to drastic measures such as error.

The Maybe type is also a monad. It is a simple kind of error monad, where all errors are represented by Nothing. A richer error monad can be built using the Either type.

Constructors

Nothing

Just a

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

Integers are stored in a kind of sign-magnitude form, hence do not expect two's complement form when using bit operations.

If the value is small (fit into an Int), IS constructor is used. Otherwise Integer and IN constructors are used to store a BigNat representing respectively the positive or the negative value magnitude.

Invariant: Integer and IN are used iff value doesn't fit in IS

class Semigroup a => Monoid a where

Plutus Tx version of Monoid.

Methods

mempty :: a

Plutus Tx version of mempty.

class Functor f => Applicative (f :: Type -> Type) where

Plutus Tx version of Applicative.

Methods

pure :: a -> f a

Plutus Tx version of pure.

(<*>) :: f (a -> b) -> f a -> f b infixl 4

Plutus Tx version of (<*>).

class Functor (f :: Type -> Type) where

Plutus Tx version of Functor.

Methods

fmap :: (a -> b) -> f a -> f b

Plutus Tx version of fmap.

class Enum a where

Class Enum defines operations on sequentially ordered types.

Methods

succ :: a -> a

The successor of a value. For numeric types, succ adds 1.

For types that implement Ord, succ x should be the least element that is greater than x, and error if there is none.

pred :: a -> a

The predecessor of a value. For numeric types, pred subtracts 1.

For types that implement Ord, pred x should be the greatest element that is less than x, and error if there is none.

toEnum :: Integer -> a

Convert from an Integer.

fromEnum :: a -> Integer

Convert to an Integer.

enumFromTo :: a -> a -> [a]

Construct a list from the given range (corresponds to [a..b]).

enumFromThenTo :: a -> a -> a -> [a]

Construct a list from the given range (corresponds to [a,b..c]). This has the same semantics as the Haskell version,so if a==b and c>=b then you get an infinite list, which you probably don't want in Plutus Core.

class Eq a => Ord a where

The Ord class is used for totally ordered datatypes.

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare :: a -> a -> Ordering

(<) :: a -> a -> Bool infix 4

(<=) :: a -> a -> Bool infix 4

(>) :: a -> a -> Bool infix 4

(>=) :: a -> a -> Bool infix 4

max :: a -> a -> a

min :: a -> a -> a

class (Ring s, AdditiveGroup v) => Module s v | v -> s where

A module, with a type of scalars which can be used to scale the values.

Methods

scale :: s -> v -> v

data Rational

Represents an arbitrary-precision ratio.

The following two invariants are maintained:

The denominator is greater than zero.

The numerator and denominator are coprime.

class Eq a where

The Eq class defines equality (==).

Methods

(==) :: a -> a -> Bool infix 4

type Semiring a = (AdditiveMonoid a, MultiplicativeMonoid a)

A semiring.

newtype Min a

Constructors

Min

Fields

getMin :: a

newtype Max a

Constructors

Max

Fields

getMax :: a

newtype Additive a

A newtype wrapper to derive Additive classes via.

Constructors

Additive a

newtype Multiplicative a

A newtype wrapper to derive Multiplicative classes via.

Constructors

Multiplicative a

newtype Join a

A wrapper witnessing that a join semi-lattice is a monoid with (\/) and bottom.

Constructors

Join a

class Monoid a => Group a where

Methods

inv :: a -> a

data BuiltinList a

data BuiltinBLS12_381_MlResult

data BuiltinBLS12_381_G2_Element

data BuiltinBLS12_381_G1_Element

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

data BuiltinPair a b

data BuiltinString

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

data BuiltinUnit

data BuiltinBool

newtype BuiltinByteStringUtf8

Constructors

BuiltinByteStringUtf8

Fields

unBuiltinByteStringUtf8 :: BuiltinByteString

type Ring a = (AdditiveGroup a, MultiplicativeMonoid a)

A ring.

class MultiplicativeSemigroup a => MultiplicativeMonoid a where

A Semigroup that it is sensible to describe using multiplication and one.

Methods

one :: a

class MultiplicativeSemigroup a where

A Semigroup that it is sensible to describe using multiplication.

Methods

(*) :: a -> a -> a infixl 7

class AdditiveMonoid a => AdditiveGroup a where

A Group that it is sensible to describe using addition, zero, and subtraction.

Methods

(-) :: a -> a -> a infixl 6

class AdditiveSemigroup a => AdditiveMonoid a where

A Monoid that it is sensible to describe using addition and zero.

Methods

zero :: a

class AdditiveSemigroup a where

A Semigroup that it is sensible to describe using addition.

Methods

(+) :: a -> a -> a infixl 6

newtype Meet a

A wrapper witnessing that a meet semi-lattice is a monoid with (/\) and top.

Constructors

Meet a

type BoundedLattice a = (BoundedJoinSemiLattice a, BoundedMeetSemiLattice a)

A bounded lattice.

class MeetSemiLattice a => BoundedMeetSemiLattice a where

A bounded meet semi-lattice, i.e. a meet semi-lattice augmented with a distinguished element top which is the unit of (/\).

Methods

top :: a

class JoinSemiLattice a => BoundedJoinSemiLattice a where

A bounded join semi-lattice, i.e. a join semi-lattice augmented with a distinguished element bottom which is the unit of (\/).

Methods

bottom :: a

type Lattice a = (JoinSemiLattice a, MeetSemiLattice a)

A lattice.

class MeetSemiLattice a where

A meet semi-lattice, i.e. a partially ordered set equipped with a binary operation (/\).

Note that the mathematical definition would require an ordering constraint - we omit that so we can define instances for e.g. (->).

Methods

(/\) :: a -> a -> a

class JoinSemiLattice a where

A join semi-lattice, i.e. a partially ordered set equipped with a binary operation (\/).

Note that the mathematical definition would require an ordering constraint - we omit that so we can define instances for e.g. (->).

Methods

(\/) :: a -> a -> a

class UnsafeFromData a where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

class FromData a where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

either :: (a -> c) -> (b -> c) -> Either a b -> c

Plutus Tx version of either

($) :: (a -> b) -> a -> b infixr 0

Plutus Tx version of 'Data.Function.($)'.

otherwise :: Bool

otherwise is defined as the value True. It helps to make guards more readable. eg.

 f x | x < 0     = ...
     | otherwise = ...

return :: Monad m => a -> m a

Inject a value into the monadic type.

(>>) :: Monad m => m a -> m b -> m b infixl 1

Sequentially compose two actions, discarding any value produced by the first, like sequencing operators (such as the semicolon) in imperative languages.

'as >> bs' can be understood as the do expression

do as
   bs

(>>=) :: Monad m => m a -> (a -> m b) -> m b infixl 1

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

'as >>= bs' can be understood as the do expression

do a <- as
   bs a

(*>) :: Applicative f => f a -> f b -> f b infixl 4

Plutus Tx version of (*>).

mconcat :: Monoid a => [a] -> a

Plutus Tx version of mconcat.

mappend :: Monoid a => a -> a -> a

Plutus Tx version of mappend.

(/=) :: Eq a => a -> a -> Bool infix 4

id :: a -> a

Plutus Tx version of id.

fst :: (a, b) -> a

Plutus Tx version of fst

snd :: (a, b) -> b

Plutus Tx version of snd

mapMaybe :: (a -> Maybe b) -> [a] -> [b]

Plutus Tx version of mapMaybe.

>>> mapMaybe (\i -> if i == 2 then Just '2' else Nothing) [1, 2, 3, 4]
"2"

liftA2 :: Applicative f => (a -> b -> c) -> f a -> f b -> f c

Plutus Tx version of liftA2.

fromMaybe :: a -> Maybe a -> a

Plutus Tx version of fromMaybe

fix :: ((a -> b) -> a -> b) -> a -> b

unless :: Applicative f => Bool -> f () -> f ()

Plutus Tx version of unless.

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

Plutus Tx version of (<$>).

negate :: AdditiveGroup a => a -> a

uncurry :: (a -> b -> c) -> (a, b) -> c

const :: a -> b -> a

Plutus Tx version of const.

even :: Integer -> Bool

odd :: Integer -> Bool

flip :: (a -> b -> c) -> b -> a -> c

Plutus Tx version of flip.

isRight :: Either a b -> Bool

Return True if the given value is a Right-value, False otherwise.

not :: Bool -> Bool

Logical negation

>>> not True
False

(<*) :: Applicative f => f a -> f b -> f a infixl 4

Plutus Tx version of (<*).

takeByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the n length prefix of a ByteString.

(=<<) :: Monad m => (a -> m b) -> m a -> m b infixr 1

Same as >>=, but with the arguments interchanged.

error :: () -> a

Aborts evaluation with an error.

trace :: BuiltinString -> a -> a

Emit the given string as a trace message before evaluating the argument.

(&&) :: Bool -> Bool -> Bool infixr 3

Logical AND. Short-circuits if the first argument evaluates to False.

>>> True && False
False

(||) :: Bool -> Bool -> Bool infixr 2

Logical OR. Short-circuits if the first argument evaluates to True.

>>> True || False
True

(<$) :: Functor f => a -> f b -> f a infixl 4

Plutus Tx version of (<$).

(.) :: (b -> c) -> (a -> b) -> a -> c infixr 9

Plutus Tx version of 'Prelude.(.)'.

until :: (a -> Bool) -> (a -> a) -> a -> a

Plutus Tx version of until.

abs :: (Ord n, AdditiveGroup n) => n -> n

Absolute value for any AdditiveGroup.

maybe :: b -> (a -> b) -> Maybe a -> b

Plutus Tx version of maybe.

>>> maybe "platypus" (\s -> s) (Just "plutus")
"plutus"
>>> maybe "platypus" (\s -> s) Nothing
"platypus"

isJust :: Maybe a -> Bool

Check if a Maybe a is Just a

>>> isJust Nothing
False
>>> isJust (Just "plutus")
True

isNothing :: Maybe a -> Bool

Check if a Maybe a is Nothing

>>> isNothing Nothing
True
>>> isNothing (Just "plutus")
False

round :: Rational -> Integer

round r returns the nearest Integer value to r. If r is equidistant between two values, the even value will be given.

quotRem :: Integer -> Integer -> (Integer, Integer)

Simultaneous quot and rem.

divMod :: Integer -> Integer -> (Integer, Integer)

Simultaneous div and mod.

curry :: ((a, b) -> c) -> a -> b -> c

(<&>) :: Functor f => f a -> (a -> b) -> f b infixl 1

Plutus Tx version of (<&>).

isLeft :: Either a b -> Bool

Return True if the given value is a Left-value, False otherwise.

divide :: Integer -> Integer -> Integer

Integer division, rounding downwards

>>> divide (-41) 5
-9

check :: Bool -> BuiltinUnit

Checks a Bool and aborts if it is false.

byteStringToInteger :: ByteOrder -> BuiltinByteString -> Integer

Convert a BuiltinByteString to a BuiltinInteger, as described in CIP-121. The first argument indicates the endianness of the conversion and the second is the bytestring to be converted. There is no limitation on the size of the bytestring. The empty bytestring is converted to the integer 0.

encodeUtf8 :: BuiltinString -> BuiltinByteString

Convert a String into a ByteString.

decodeUtf8 :: BuiltinByteString -> BuiltinString

Converts a ByteString to a String.

sha2_256 :: BuiltinByteString -> BuiltinByteString

The SHA2-256 hash of a ByteString

sha3_256 :: BuiltinByteString -> BuiltinByteString

The SHA3-256 hash of a ByteString

blake2b_224 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-224 hash of a ByteString

blake2b_256 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-256 hash of a ByteString

keccak_256 :: BuiltinByteString -> BuiltinByteString

The KECCAK-256 hash of a ByteString

ripemd_160 :: BuiltinByteString -> BuiltinByteString

The RIPEMD-160 hash of a ByteString

verifyEcdsaSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (33 bytes)

-> BuiltinByteString
Message hash (32 bytes)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given an ECDSA SECP256k1 verification key, an ECDSA SECP256k1 signature, and an ECDSA SECP256k1 message hash (all as BuiltinByteStrings), verify the hash with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length:

The first byte of the public key must correspond to the sign of the y coordinate: this is 0x02 if y is even, and 0x03 otherwise.

The remaining bytes of the public key must correspond to the x coordinate, as a big-endian integer.

The first 32 bytes of the signature must correspond to the big-endian integer representation of _r_.

The last 32 bytes of the signature must correspond to the big-endian integer representation of _s_.

While this primitive accepts a hash, any caller should only pass it hashes that they computed themselves: specifically, they should receive the message from a sender and hash it, rather than receiving the hash from said sender. Failure to do so can be dangerous. Other than length, we make no requirements of what hash gets used.

See also

secp256k1_ec_pubkey_serialize; this implements the format for the verification key that we accept, given a length argument of 33 and the SECP256K1_EC_COMPRESSED flag.

secp256k1_ecdsa_serialize_compact; this implements the format for the signature that we accept.

verifySchnorrSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (32 bytes)

-> BuiltinByteString
Message (arbitrary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given a Schnorr SECP256k1 verification key, a Schnorr SECP256k1 signature, and a message (all as BuiltinByteStrings), verify the message with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length. Throughout, we refer to co-ordinates of the point R.

The bytes of the public key must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The first 32 bytes of the signature must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The last 32 bytes of the signature must correspond to the bytes of s, as a big-endian integer, as specified in BIP-340.

See also

BIP-340

secp256k1_xonly_pubkey_serialize; this implements the format for the verification key that we accept.

secp256k1_schnorrsig_sign; this implements the signing logic for signatures this builtin can verify.

verifyEd25519Signature

Arguments

:: BuiltinByteString
Public Key (32 bytes)

-> BuiltinByteString
Message (arbirtary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Ed25519 signature verification. Verify that the signature is a signature of the message by the public key. This will fail if key or the signature are not of the expected length.

expMod :: Integer -> Integer -> Integer -> Integer

FIXME

integerToByteString :: ByteOrder -> Integer -> Integer -> BuiltinByteString

Convert a BuiltinInteger into a BuiltinByteString, as described in CIP-121. The first argument indicates the endianness of the conversion and the third argument is the integer to be converted, which must be non-negative. The second argument must also be non-negative and it indicates the required width of the output. If the width is zero then the output is the smallest bytestring which can contain the converted input (and in this case, the integer 0 encodes to the empty bytestring). If the width is nonzero then the output bytestring will be padded to the required width with 0x00 bytes (on the left for big-endian conversions and on the right for little-endian conversions); if the input integer is too big to fit into a bytestring of the specified width then the conversion will fail. Conversion will also fail if the specified width is greater than 8192 or the input integer is too big to fit into a bytestring of length 8192.

andByteString :: Bool -> BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Perform logical AND on two BuiltinByteString arguments, as described in CIP-122.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

orByteString :: Bool -> BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Perform logical OR on two BuiltinByteString arguments, as described here.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

xorByteString :: Bool -> BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Perform logical XOR on two BuiltinByteString arguments, as described here.

The first argument indicates whether padding semantics should be used or not; if False, truncation semantics will be used instead.

See also

Padding and truncation semantics

Bit indexing scheme

complementByteString :: BuiltinByteString -> BuiltinByteString

Perform logical complement on a BuiltinByteString, as described here.

See also

Bit indexing scheme

readBit :: BuiltinByteString -> Integer -> Bool

Read a bit at the _bit_ index given by the Integer argument in the BuiltinByteString argument. The result will be True if the corresponding bit is set, and False if it is clear. Will error if given an out-of-bounds index argument; that is, if the index is either negative, or equal to or greater than the total number of bits in the BuiltinByteString argument.

See also

Bit indexing scheme

Operation description

writeBits :: BuiltinByteString -> [Integer] -> Bool -> BuiltinByteString

Given a BuiltinByteString, a list of indexes to change, and a boolean value b to change those indexes to, set the bit at each of the specified index as follows:

If b is True, set that bit;

Otherwise, clear that bit.

Will error if any of the indexes are out-of-bounds: that is, if the index is either negative, or equal to or greater than the total number of bits in the BuiltinByteString argument.

Note

This differs slightly from the description of the corresponding operation in CIP-122; instead of a single changelist argument comprised of pairs, we instead pass a single list of indexes to change, and a single boolean value to change those indexes to. The original proposal allowed one to set and clear bits in a single operation, but constructing the list of boolean values for the updates was somewhat expensive. If it's really necessary to set some bits and clear others then it is easier to call the function twice, once to set bits and and once to clear them.

See also

Bit indexing scheme

Operation description

replicateByte :: Integer -> Integer -> BuiltinByteString

Given a length (first argument) and a byte (second argument), produce a BuiltinByteString of that length, with that byte in every position. Will error if given a negative length, or a second argument that isn't a byte (less than 0, greater than 255).

See also

Operation description

shiftByteString :: BuiltinByteString -> Integer -> BuiltinByteString

Shift a BuiltinByteString, as per CIP-123.

rotateByteString :: BuiltinByteString -> Integer -> BuiltinByteString

Rotate a BuiltinByteString, as per CIP-123.

countSetBits :: BuiltinByteString -> Integer

Count the set bits in a BuiltinByteString, as per CIP-123.

findFirstSetBit :: BuiltinByteString -> Integer

Find the lowest index of a set bit in a BuiltinByteString, as per CIP-123.

If given a BuiltinByteString which consists only of zero bytes (including the empty BuiltinByteString, this returns -1.

appendByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Concatenates two ByteStrings.

consByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Adds a byte to the front of a ByteString.

sliceByteString :: Integer -> Integer -> BuiltinByteString -> BuiltinByteString

Returns the substring of a ByteString from index start of length n.

lengthOfByteString :: BuiltinByteString -> Integer

Returns the length of a ByteString.

indexByteString :: BuiltinByteString -> Integer -> Integer

Returns the byte of a ByteString at index.

emptyByteString :: BuiltinByteString

An empty ByteString.

appendString :: BuiltinString -> BuiltinString -> BuiltinString

Append two Strings.

emptyString :: BuiltinString

An empty String.

equalsString :: BuiltinString -> BuiltinString -> Bool

Check if two strings are equal

bls12_381_G1_equals :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> Bool

bls12_381_G1_add :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_neg :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_scalarMul :: Integer -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compress :: BuiltinBLS12_381_G1_Element -> BuiltinByteString

bls12_381_G1_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compressed_zero :: BuiltinByteString

bls12_381_G1_compressed_generator :: BuiltinByteString

bls12_381_G2_equals :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> Bool

bls12_381_G2_add :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_neg :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_scalarMul :: Integer -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compress :: BuiltinBLS12_381_G2_Element -> BuiltinByteString

bls12_381_G2_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compressed_zero :: BuiltinByteString

bls12_381_G2_compressed_generator :: BuiltinByteString

bls12_381_millerLoop :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_MlResult

bls12_381_mulMlResult :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult

bls12_381_finalVerify :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> Bool

fromOpaque :: HasFromOpaque arep a => arep -> a

toOpaque :: HasToOpaque a arep => a -> arep

fromBuiltin :: HasFromBuiltin arep => arep -> FromBuiltin arep

toBuiltin :: HasToBuiltin a => a -> ToBuiltin a

traceError :: BuiltinString -> a

Log a message and then terminate the evaluation with an error.

traceIfFalse :: BuiltinString -> Bool -> Bool

Emit the given BuiltinString only if the argument evaluates to False.

traceIfTrue :: BuiltinString -> Bool -> Bool

Emit the given BuiltinString only if the argument evaluates to True.

traceBool :: BuiltinString -> BuiltinString -> Bool -> Bool

Emit one of two BuiltinString depending on whether or not the argument evaluates to True or False.

gsub :: Group a => a -> a -> a

unsafeRatio :: Integer -> Integer -> Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

ratio :: Integer -> Integer -> Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

modulo :: Integer -> Integer -> Integer

Integer remainder, always positive for a positive divisor

>>> modulo (-41) 5
4

quotient :: Integer -> Integer -> Integer

Integer division, rouding towards zero

>>> quotient (-41) 5
-8

remainder :: Integer -> Integer -> Integer

Integer remainder, same sign as dividend

>>> remainder (-41) 5
-1

dropByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the suffix of a ByteString after n elements.

module PlutusTx.List

sum :: (Foldable t, AdditiveMonoid a) => t a -> a

Plutus Tx version of sum.

class IsString a where

IsString is used in combination with the -XOverloadedStrings language extension to convert the literals to different string types.

For example, if you use the text package, you can say

{-# LANGUAGE OverloadedStrings  #-}

myText = "hello world" :: Text

Internally, the extension will convert this to the equivalent of

myText = fromString @Text ("hello world" :: String)

Note: You can use fromString in normal code as well, but the usual performance/memory efficiency problems with String apply.

Methods

fromString :: String -> a

class ToData a where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

class FromData a where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData a where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

divideInteger :: Integer -> Integer -> Integer

Divide two integers.

ifThenElse :: forall (a :: Type). Bool -> a -> a -> a

fromInteger :: Num a => Integer -> a

Conversion from an Integer. An integer literal represents the application of the function fromInteger to the appropriate value of type Integer, so such literals have type (Num a) => a.

Helpers for Data encoding

ToData

productToData2 :: forall (a :: Type) (b :: Type). (ToData a, ToData b) => a -> b -> BuiltinData

Helper to write toBuiltinData for 2-products (something isomorphic to a pair).

Notes

This is necessary because, by default, the TH-driven Plutus derivation for a type such as this:

data Foo = Foo Integer Integer

Will encode as Constr 0. This is inefficient: we only have one 'arm', and the tag will always be the same for any valid value, but we have to carry it around anyway. This function instead encodes it as a list.

It is essential that this only be used together with productFromData2 (or its ticked equivalent) and productUnsafeFromData2; otherwise, the encodings may not roundtrip.

productFromData2 will verify that:

It is given a Plutus list;

That this list has exactly two elements; and

Each element of this list decodes to the correct type.

productUnsafeFromData2 will instead decode without error when:

It is given a Plutus list;

That this list has at least two elements; and

Each element of this list decodes to the correct type.

More precisely, productFromData2 verifies that it is given exactly the right amount of data, whereas productUnsafeFromData2 works when it is given /at least/ the right amount of data (but can be given more). This implies that if productFromData2 f dat = Just x, then productUnsafeFromData2 f dat = x, but that the converse does not necessarily hold. This (arguably unusual) behaviour matches the definitions generated by the Plutus-provided TH.

How to use this

data Foo = Foo Integer Integer

instance ToData Foo where
  {-# INLINEABLE toBuiltinData #-}
  toBuiltinData (Foo x y) = productToData2 x y

instance FromData Foo where
  {-# INLINEABLE fromBuiltinData #-}
  fromBuiltinData = productFromData2 Foo

instance UnsafeFromData Foo where
  {-# INLINEABLE unsafeFromBuiltinData #-}
  unsafeFromBuiltinData = productUnsafeFromData2 Foo

productToData3 :: forall (a :: Type) (b :: Type) (c :: Type). (ToData a, ToData b, ToData c) => a -> b -> c -> BuiltinData

As productToData2, but for 3-products.

productToData4 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type). (ToData a, ToData b, ToData c, ToData d) => a -> b -> c -> d -> BuiltinData

As productToData2, but for 4-products.

productToData5 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type). (ToData a, ToData b, ToData c, ToData d, ToData e) => a -> b -> c -> d -> e -> BuiltinData

As productToData2, but for 5-products.

productToData6 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type). (ToData a, ToData b, ToData c, ToData d, ToData e, ToData f) => a -> b -> c -> d -> e -> f -> BuiltinData

As productToData2, but for 6-products.

FromData

productFromData2 :: forall (a :: Type) (b :: Type) (c :: Type). (FromData a, FromData b) => (a -> b -> c) -> BuiltinData -> Maybe c

Helper to write fromBuiltinData for 2-products (something isomorphic to a pair).

Note

See productToData2 for examples of use and why this is necessary.

Use this function if you don't need any verification beyond the components of the product decoding successfully.

productFromData2' :: forall (a :: Type) (b :: Type) (c :: Type). (FromData a, FromData b) => (a -> b -> Maybe c) -> BuiltinData -> Maybe c

As productFromData2, but allows additional checks beyond the components successfully decoding.

productFromData3 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type). (FromData a, FromData b, FromData c) => (a -> b -> c -> d) -> BuiltinData -> Maybe d

As productFromData2, but for 3-products.

productFromData3' :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type). (FromData a, FromData b, FromData c) => (a -> b -> c -> Maybe d) -> BuiltinData -> Maybe d

As productFromData3, but allows additional checks beyond the components successfully decoding.

productFromData4 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type). (FromData a, FromData b, FromData c, FromData d) => (a -> b -> c -> d -> e) -> BuiltinData -> Maybe e

As productFromData2, but for 4-products.

productFromData4' :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type). (FromData a, FromData b, FromData c, FromData d) => (a -> b -> c -> d -> Maybe e) -> BuiltinData -> Maybe e

As productFromData4, but allows additional checks beyond the components successfully decoding.

productFromData5 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type). (FromData a, FromData b, FromData c, FromData d, FromData e) => (a -> b -> c -> d -> e -> f) -> BuiltinData -> Maybe f

As productFromData2, but for 5-products.

productFromData5' :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type). (FromData a, FromData b, FromData c, FromData d, FromData e) => (a -> b -> c -> d -> e -> Maybe f) -> BuiltinData -> Maybe f

As productFromData5, but allows additional checks beyond the components successfully decoding.

productFromData6 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type) (g :: Type). (FromData a, FromData b, FromData c, FromData d, FromData e, FromData f) => (a -> b -> c -> d -> e -> f -> g) -> BuiltinData -> Maybe g

As productFromData2, but for 6-products.

productFromData6' :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type) (g :: Type). (FromData a, FromData b, FromData c, FromData d, FromData e, FromData f) => (a -> b -> c -> d -> e -> f -> Maybe g) -> BuiltinData -> Maybe g

As productFromData6, but allows additional checks beyond the components successfully decoding.

UnsafeFromData

productUnsafeFromData2 :: forall (a :: Type) (b :: Type) (c :: Type). (UnsafeFromData a, UnsafeFromData b) => (a -> b -> c) -> BuiltinData -> c

Helper to write unsafeFromBuiltinData for 2-products (something isomorphic to a pair).

Note

See productToData2 for examples of use, and why this is necessary.

productUnsafeFromData3 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type). (UnsafeFromData a, UnsafeFromData b, UnsafeFromData c) => (a -> b -> c -> d) -> BuiltinData -> d

As productUnsafeFromData2, but for 3-products.

productUnsafeFromData4 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type). (UnsafeFromData a, UnsafeFromData b, UnsafeFromData c, UnsafeFromData d) => (a -> b -> c -> d -> e) -> BuiltinData -> e

As productUnsafeFromData2, but for 4-products.

productUnsafeFromData5 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type). (UnsafeFromData a, UnsafeFromData b, UnsafeFromData c, UnsafeFromData d, UnsafeFromData e) => (a -> b -> c -> d -> e -> f) -> BuiltinData -> f

As productUnsafeFromData2, but for 5-products.

productUnsafeFromData6 :: forall (a :: Type) (b :: Type) (c :: Type) (d :: Type) (e :: Type) (f :: Type) (g :: Type). (UnsafeFromData a, UnsafeFromData b, UnsafeFromData c, UnsafeFromData d, UnsafeFromData e, UnsafeFromData f) => (a -> b -> c -> d -> e -> f -> g) -> BuiltinData -> g

As productUnsafeFromData2, but for 6-products.