TrustlessSidechain.Reserve

Documentation

mkReserveValidator :: VersionOracleConfig -> BuiltinData -> ReserveRedeemer -> ScriptContext -> Bool

Error codes description follows:

ERROR-RESERVE-01: Governance approval is not present ERROR-RESERVE-02: Other tokens than governance token are minted or burnt ERROR-RESERVE-03: Datum of a propagated reserve utxo changes ERROR-RESERVE-04: Assets of a propagated reserve utxo don't increase by reserve tokens ERROR-RESERVE-05: No unique input utxo carrying authentication token ERROR-RESERVE-06: No unique output utxo at the reserve address and carrying authentication token ERROR-RESERVE-07: Datum of input reserve utxo malformed ERROR-RESERVE-08: Datum of output reserve utxo malformed ERROR-RESERVE-09: Datum of a propagated reserve utxo changes not only by immutable settings ERROR-RESERVE-10: Assets of a propagated reserve utxo change ERROR-RESERVE-11: V(t) tokens are not minted ERROR-RESERVE-12: Other tokens than V(t) tokens are minted or burnt ERROR-RESERVE-13: Assets of a propagated reserve utxo don't decrease by reserve tokens in desired way ERROR-RESERVE-14: Datum of a propagated reserve utxo changes not only by stats in desired way ERROR-RESERVE-15: Incorrect amount of reserve tokens goes into an illiquid circulation supply ERROR-RESERVE-16: No unique output utxo at the illiquid circulation supply address ERROR-RESERVE-17: An authentication token is not burnt ERROR-RESERVE-18: Other tokens than auth token and governance token are minted or burnt ERROR-RESERVE-19: Not all reserve tokens are transferred to illiquid circulation supply ERROR-RESERVE-20: Reserve utxo input exists without an authentication token ERROR-RESERVE-21: Reserve utxo output exists without an authentication token ERROR-RESERVE-22: Governance approval is not present ERROR-RESERVE-23: Other tokens than governance token are minted or burnt ERROR-RESERVE-24: Governance approval is not present ERROR-RESERVE-25: Continuing output exists without an authentication token

mkReserveValidatorUntyped :: BuiltinData -> BuiltinData -> BuiltinData -> BuiltinData -> BuiltinUnit

serialisableReserveValidator :: SerialisedScript

mkReserveAuthPolicy :: VersionOracleConfig -> BuiltinData -> ScriptContext -> Bool

Error codes description follows:

ERROR-RESERVE-AUTH-01: Governance approval is not present ERROR-RESERVE-AUTH-02: Single reserve authentication token is not minted ERROR-RESERVE-AUTH-03: Output reserve UTxO doesn't carry auth token ERROR-RESERVE-AUTH-04: Output reserve UTxO doesn't carry correct initial datum ERROR-RESERVE-AUTH-05: Output reserve UTxO carries other tokens ERROR-RESERVE-AUTH-06: No unique output UTxO at the reserve address ERROR-RESERVE-AUTH-07: Output reserve UTxO carries no inline datum or malformed datum

mkReserveAuthPolicyUntyped :: BuiltinData -> BuiltinData -> BuiltinData -> BuiltinUnit

serialisableReserveAuthPolicy :: SerialisedScript

reserveAuthTokenTokenName :: TokenName