TrustlessSidechain.Types

Documentation

data PermissionedCandidatesPolicyRedeemer

PermissionedCandidatesPolicyRedeemer signals whether transaction is supposed to mint or burn PermissionedCandidates tokens

Constructors

PermissionedCandidatesMint

PermissionedCandidatesBurn

data PermissionedCandidatesValidatorRedeemer

PermissionedCandidatesValidatorRedeemer signals whether transaction is supposed to update the list of permissioned candidates or remove the list altogether.

Constructors

UpdatePermissionedCandidates

RemovePermissionedCandidates

data ImmutableReserveSettings

Constructors

ImmutableReserveSettings

Fields

tokenKind :: AssetClass

tokenKind is an asset class of tokens that a reserve UTxO is allowed to store

data MutableReserveSettings

Constructors

MutableReserveSettings

Fields

vFunctionTotalAccrued :: CurrencySymbol

vFunctionTotalAccrued is a currency symbol of a minting policy that dictates the upper bound on the number of tokenKind tokens that can be transferred from a reserve utxo to an illiquid circulation supply from t0 till now

incentiveAmount :: Integer

The amount of tokenKind the user is allowed to claim when releasing money from the reserve

newtype ReserveStats

Constructors

ReserveStats

Fields

tokenTotalAmountTransferred :: Integer

tokenTotalAmountTransferred is the total number of tokens that already have been transferred from a reserve utxo to an illiquid circulation supply

data ReserveDatum

Constructors

ReserveDatum

Fields

immutableSettings :: ImmutableReserveSettings

mutableSettings :: MutableReserveSettings

stats :: ReserveStats

data ReserveRedeemer

Constructors

DepositToReserve

TransferToIlliquidCirculationSupply

UpdateReserve

Handover

data IlliquidCirculationSupplyRedeemer

Constructors

DepositMoreToSupply

WithdrawFromSupply

data VersionedGenericDatum a

Constructors

VersionedGenericDatum

Fields

datum :: a

genericData :: BuiltinData

version :: Integer

newtype VersionOracle

Datum attached to VersionOraclePolicy tokens stored on the VersionOracleValidator script.

Constructors

VersionOracle

Fields

scriptId :: Integer

Unique identifier of the validator.

newtype VersionOracleConfig

Configuration of the versioning system. Contains currency symbol of VersionOraclePolicy tokens. Required to identify versioning tokens that can be trusted.

Constructors

VersionOracleConfig

Fields

versionOracleCurrencySymbol :: CurrencySymbol

@since v5.0.0

data VersionOracleDatum

Datum attached to VersionOraclePolicy tokens stored on the VersionOracleValidator script.

Constructors

VersionOracleDatum

Fields

versionOracle :: VersionOracle

VersionOracle which identifies the script.

currencySymbol :: CurrencySymbol

Currency Symbol of the VersioningOraclePolicy tokens.

data VersionOraclePolicyRedeemer

Redeemer for the versioning oracle minting policy that instructs the script whether to mint or burn versioning tokens.

Constructors

InitializeVersionOracle VersionOracle ScriptHash
Mint versioning tokens from init tokens. Used during sidechain initialization.

MintVersionOracle VersionOracle ScriptHash
Mint a new versioning token ensuring it contains correct datum and reference script.

BurnVersionOracle VersionOracle
Burn existing versioning token.